third_party/blink/web_tests/external/wpt/WebCryptoAPI/generateKey/failures_Ed448.https.any.worker-expected.txt

This is a testharness.js-based test.
Found 26 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Bad usages: generateKey({name: Ed448}, true, [encrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, encrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [verify, sign, encrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, verify, sign, sign, verify, encrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [decrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, decrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [verify, sign, decrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, verify, sign, sign, verify, decrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [wrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, wrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [verify, sign, wrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, verify, sign, sign, verify, wrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [unwrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, unwrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [verify, sign, unwrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, verify, sign, sign, verify, unwrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [deriveKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, deriveKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [verify, sign, deriveKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, verify, sign, sign, verify, deriveKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [deriveBits])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, deriveBits])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [verify, sign, deriveBits])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: Ed448}, true, [sign, verify, sign, sign, verify, deriveBits])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Empty usages: generateKey({name: Ed448}, false, [])
  assert_equals: Empty usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Empty usages: generateKey({name: Ed448}, true, [])
  assert_equals: Empty usages not supported expected "SyntaxError" but got "NotSupportedError"
Harness: the test ran to completion.
